requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiogram==2.12.1
    # via -r requirements.in
aiohttp==3.7.4.post0
    # via aiogram
aiosqlite==0.16.1
    # via
    #   -r requirements.in
    #   tortoise-orm
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
babel==2.9.1
    # via aiogram
certifi==2021.5.30
    # via aiogram
chardet==4.0.0
    # via aiohttp
idna==3.2
    # via yarl
iso8601==0.1.14
    # via tortoise-orm
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
pypika-tortoise==0.1.1
    # via tortoise-orm
python-dotenv==0.17.1
    # via -r requirements.in
pytz==2021.1
    # via
    #   babel
    #   tortoise-orm
tortoise-orm==0.17.3
    # via -r requirements.in
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   aiosqlite
yarl==1.6.3
    # via aiohttp
